Split out ID handling.
